embeddings/src/requirements.txt
python-dotenv
supabase
transformers
bottle
InstructorEmbedding
torchvision
sentence-transformers
sacremoses